fix: [ANDROAPP-7255] Correct quick login with biometrics flow [skip size] (#4430)
* fix: [ANDROAPP-7255] Correct quick login with biometrics flow

* fix: [ANDROAPP-7255] ktlint

* fix: [ANDROAPP-7255] update default return value for biometric permission

* fix: [ANDROAPP-7255] use main thread to login with biometrics

* fix: [ANDROAPP-7255] fix crash when cancelling login job with biometrics, do not show biometrics dialog after login out

* fix: [ANDROAPP-7255] fix unit tests

* fix: [ANDROAPP-7255] ktlint

* fix: review comments

* fix: [ANDROAPP-7255] Address code review comments

* fix: [ANDROAPP-7255] Address copilot code review comments

* fix: [ANDROAPP-7255] make OidcInfo nullable in login module

---------

+++ app/src/main/java/org/dhis2/usescases/main/MainPresenter.kt
@@ -250,6 +250,7 @@ class MainPresenter(
     fun onDeleteAccount() {
         view.showProgressDeleteNotification()
         try {
+            repository.checkDeleteBiometricsPermission()
             workManagerController.cancelAllWork()
             syncStatusController.restore()
             deleteUserData.wipeCacheAndPreferences(view.obtainFileView())
